refactor: move agent_config from domain to orchestration namespace
agent_config was under terok.lib.domain but classified as
orchestration-layer in tach — a known oddity that import-linter
flagged as a cross-layer violation. Moving it to its true home
eliminates the ignore_imports exception.
