+++ .importlinter
@@ -17,11 +17,9 @@ layers =
     terok.lib.domain
     terok.lib.orchestration
     terok.lib.core
-# TODO: resolve these layer violations instead of ignoring them:
-# - agent_config lives under domain namespace but acts as orchestration — move to orchestration
+# TODO: resolve this layer violation instead of ignoring it:
 # - clearance CLI command delegates to TUI screen — extract shared logic to domain/lib
 ignore_imports =
-    terok.lib.orchestration.task_runners -> terok.lib.domain.agent_config
     terok.cli.commands.clearance -> terok.tui.clearance_screen
 
 # ── Cross-package boundaries (tach can't do this) ────────────────
@@ -70,9 +68,9 @@ allowed_importers =
     terok.cli.main
     terok.cli.wiring
     terok.lib.core.projects
-    terok.lib.domain.agent_config
     terok.lib.domain.facade
     terok.lib.domain.project
+    terok.lib.orchestration.agent_config
     terok.lib.orchestration.container_doctor
     terok.lib.orchestration.docker
     terok.lib.orchestration.environment
