+++ src/terok/cli/commands/info.py
@@ -78,7 +78,7 @@ def _cmd_config_show(project_id: str, preset: str | None) -> None:
     import json
 
     from ...lib.core.projects import load_project
-    from ...lib.domain.agent_config import build_agent_config_stack
+    from ...lib.orchestration.agent_config import build_agent_config_stack
 
     color_enabled = _supports_color()
